[Feat] 조인 결과 반환을 위한 통합 cartDTO, 엔티티 타입 수정 #81

+++ backend/src/main/java/com/synergy/backend/orders/model/request/AddCartReq.java
@@ -14,12 +14,12 @@ import java.util.List;
 public class AddCartReq {
 
     private Long productIdx;
-    private Long count;
+    private Integer count;
     private Integer price;
     private List<AddCartOption> addCartOptions;
 
     @Builder
-    public AddCartReq(Long productIdx, Long count, Integer price) {
+    public AddCartReq(Long productIdx, Integer count, Integer price) {
         this.productIdx = productIdx;
         this.count = count;
         this.price = price;
